Add tokenizer v1.0 and fix compile errors
Expereminted with getline for the tokenizer. This method will probably
not work since it throws out the delimiter. Likely I will need a method
that reads 1 char at a time and checks if its a space, (, or ) and
handles it that way.
Added guards to prevent compilation errors in expression.hpp. Should do
this for all headers.
Removed function from ASTNode that shouldn't have existed.