Update ASTBuilder to v1.1
ASTNodes are now referenced as pointers. Parent nodes are returned like
this as well in getParent, which fixed the bug in v1.0 where it would
simply return itself.
Note that this version does not contain any error checking. The returned
tree will be in a bad state if there is a syntax error in the tokenized
list.